Refactor Profile component structure
Split the Profile component in `src/pages/Profile.tsx` into smaller, manageable files while ensuring all existing functionality remains intact. Removed any unused imports and files as part of the cleanup process.
